Fixed faulty usage docs in --help
Clap always puts subcommand arguments last. Because we are abusing
the way subcommands work, and because we have an argument which is
a list of strings, this resulted in faulty help documentation. It
showed the subcommand as coming after the command argument, which is
not possible (because the command argument is a list, and so must be
last).

Used a hack involving writing two clap parsers to work around this
issue. The help documentation now doesn't reflect that you can use
`attempt` without specifying a backoff strategy, but at least that's
failing to report complete information rather than reporting faulty
information.